Move session_start + tests under hook_daemon/
Largest move: session_start.py plus test_session_start.py and
test_session_start_podman.py. Templates remain at devinfra/claude/templates/
(referenced via exports_files). Updates AGENTS.md path reference.
