+++ devinfra/claude/BUILD.bazel
@@ -1,6 +1,11 @@
 load("@rules_python//python:defs.bzl", "py_binary", "py_library")
 load("//devinfra/testing:defs.bzl", "py_test")
 
+exports_files(
+    ["templates/session_context.mako"],
+    visibility = ["//devinfra/claude/hook_daemon:__pkg__"],
+)
+
 # === Base modules (no internal deps) ===
 
 py_library(
@@ -279,41 +284,6 @@ py_library(
     ],
 )
 
-py_library(
-    name = "session_start_lib",
-    srcs = ["session_start.py"],
-    data = ["templates/session_context.mako"],
-    imports = ["../.."],
-    visibility = ["//visibility:public"],
-    deps = [
-        ":bazel_wrapper_lib",
-        ":bazelisk_setup",
-        ":buildbuddy_setup",
-        ":cli_tools_setup",
-        ":container_runtime",
-        ":debug",
-        ":env_file",
-        ":errors",
-        ":fork_remote_setup",
-        ":hook_config",
-        ":k8s_secrets_setup",
-        ":mkcert_setup",
-        ":nix_setup",
-        ":precommit_setup",
-        ":session_paths",
-        ":settings",
-        ":tmpfs_setup",
-        ":tracing",
-        "//devinfra:build_info",
-        "//devinfra/claude/auth_proxy:setup",
-        "//devinfra/claude/claude_api/hooks:session_start",
-        "//devinfra/claude/supervisor:setup",
-        "//util:env",
-        "@pypi//mako",
-        "@pypi//opentelemetry_api",
-    ],
-)
-
 # === Binaries ===
 
 py_binary(
@@ -391,92 +361,6 @@ py_test(
         "//devinfra/claude/claude_api/hooks:common",
         "//devinfra/claude/claude_api/hooks:session_start",
         "@pypi//pytest",
-        "@pypi//pytest_bazel",
-    ],
-)
-
-py_test(
-    name = "test_session_start",
-    size = "medium",
-    srcs = ["test_session_start.py"],
-    data = [
-        ":hook_dispatch",
-        "//devinfra/claude/auth_proxy:run",
-        "//devinfra/claude/testdata/test_workspace:workspace_files",
-        "//third_party/jdk:cacerts",
-        "//third_party/jdk:keytool",
-    ],
-    env = {
-        "JAVA_CACERTS_RLOCATION": "$(rlocationpath //third_party/jdk:cacerts)",
-        "KEYTOOL_RLOCATION": "$(rlocationpath //third_party/jdk:keytool)",
-    },
-    # Inherit proxy env vars so MockEgressProxy can chain through real proxy
-    env_inherit = [
-        "CURL_CA_BUNDLE",
-        "HTTPS_PROXY",
-        "https_proxy",
-        "NODE_EXTRA_CA_CERTS",
-        "REQUESTS_CA_BUNDLE",
-        "SSL_CERT_FILE",
-    ],
-    imports = ["../.."],
-    # local: keytool (Adoptium JDK) needs libjli.so which isn't on the RBE worker image
-    local = True,
-    tags = [
-        "e2e",
-        "external",
-    ],
-    deps = [
-        ":conftest",
-        ":settings",
-        "//devinfra/claude/testing:fixtures",
-        "//devinfra/claude/testing:session_start_helpers",
-        "//devinfra/claude/testing:shell_helpers",
-        "//util/testing:undeclared_outputs",
-        "@pypi//pytest",
-        "@pypi//pytest_asyncio",
-        "@pypi//pytest_bazel",
-    ],
-)
-
-py_test(
-    name = "test_session_start_podman",
-    size = "medium",
-    srcs = ["test_session_start_podman.py"],
-    data = [
-        ":hook_dispatch",
-        "//devinfra/claude/auth_proxy:run",
-        "//third_party/jdk:cacerts",
-        "//third_party/jdk:keytool",
-    ],
-    env = {
-        "JAVA_CACERTS_RLOCATION": "$(rlocationpath //third_party/jdk:cacerts)",
-        "KEYTOOL_RLOCATION": "$(rlocationpath //third_party/jdk:keytool)",
-    },
-    # Inherit proxy env vars so MockEgressProxy can chain through real proxy
-    env_inherit = [
-        "CURL_CA_BUNDLE",
-        "HTTPS_PROXY",
-        "https_proxy",
-        "NODE_EXTRA_CA_CERTS",
-        "REQUESTS_CA_BUNDLE",
-        "SSL_CERT_FILE",
-    ],
-    imports = ["../.."],
-    # local: podman needs mount/networking capabilities blocked by linux-sandbox and RBE
-    local = True,
-    tags = [
-        "e2e",
-        "external",
-        "manual",  # podman path not used in prod currently
-    ],
-    deps = [
-        ":conftest",
-        "//devinfra/claude/testing:fixtures",
-        "//devinfra/claude/testing:session_start_helpers",
-        "//devinfra/claude/testing:shell_helpers",
-        "@pypi//pytest",
-        "@pypi//pytest_asyncio",
         "@pypi//pytest_bazel",
     ],
 )
